requirements-dev.txt
pyqt5-stubs
ruff
pytest
pre-commit
pytest-qgis
